move MergeSortedN tests to correct file
