+++ akka-stream-tests/src/test/scala/akka/stream/scaladsl/GraphMergeSortedSpec.scala
@@ -4,15 +4,12 @@
 
 package akka.stream.scaladsl
 
-import akka.NotUsed
 import akka.stream._
 import akka.stream.testkit.TwoStreamsSetup
 import com.github.ghik.silencer.silent
-import org.scalacheck.{ Gen, Shrink }
+import org.scalacheck.Gen
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalacheck.Shrink
-
-import scala.collection.immutable
 
 @silent // tests deprecated apis
 class GraphMergeSortedSpec extends TwoStreamsSetup with GeneratorDrivenPropertyChecks {
@@ -44,38 +41,6 @@ class GraphMergeSortedSpec extends TwoStreamsSetup with GeneratorDrivenPropertyC
           .runWith(Sink.head)
           .futureValue should ===(0 until N)
       }
-    }
-
-    "work as a special case of MergeSortedN with N=2" in {
-      val gen = Gen.listOf(Gen.oneOf(false, true))
-
-      forAll(gen) { picks ⇒
-        val N = picks.size
-        val (left, right) = picks.zipWithIndex.partition(_._1)
-        Source.mergeSortedN(List(Source(left.map(_._2)), Source(right.map(_._2))))
-          .grouped(N max 1)
-          .concat(Source.single(Nil))
-          .runWith(Sink.head)
-          .futureValue should ===(0 until N)
-      }
-    }
-
-    "work" in {
-      // List(false, false, false, true, false, true, false, true, true, false)
-      val leftSource: Source[Int, NotUsed] = Source(List(0, 2))
-      val rightSource: Source[Int, NotUsed] = Source(List(1))
-
-      val underTest: Source[Int, NotUsed] = Source.mergeSortedN(List(leftSource, rightSource))
-
-      val grouped: Source[immutable.Seq[Int], NotUsed] = underTest
-        .grouped(4)
-
-      val sortedResult: immutable.Seq[Int] = grouped
-        .concat(Source.single(Nil))
-        .runWith(Sink.head)
-        .futureValue
-
-      sortedResult shouldBe List(0, 1, 2)
     }
 
     commonTests()
